requirements.txt
python-telegram-bot
   flask
   sqlite3